Moving packages and some comment improvements.
