+++ hub-detect/src/main/groovy/com/blackducksoftware/integration/hub/detect/workflow/diagnostic/profiling/BomToolTimekeeper.java
@@ -1,4 +1,4 @@
-package com.blackducksoftware.integration.hub.detect.workflow.diagnostic;
+package com.blackducksoftware.integration.hub.detect.workflow.diagnostic.profiling;
 
 import java.util.ArrayList;
 import java.util.HashMap;
