+++ hub-detect/src/main/groovy/com/blackducksoftware/integration/hub/detect/workflow/search/rules/BomToolSearchProvider.java
@@ -26,7 +26,7 @@ package com.blackducksoftware.integration.hub.detect.workflow.search.rules;
 import com.blackducksoftware.integration.hub.detect.bomtool.BomToolEnvironment;
 import com.blackducksoftware.integration.hub.detect.bomtool.BomToolType;
 import com.blackducksoftware.integration.hub.detect.factory.BomToolFactory;
-import com.blackducksoftware.integration.hub.detect.workflow.diagnostic.BomToolProfiler;
+import com.blackducksoftware.integration.hub.detect.workflow.diagnostic.profiling.BomToolProfiler;
 
 public class BomToolSearchProvider {
     private final BomToolFactory bomToolFactory;
